requirements.txt
streamlit>=1.32.0
openai>=1.12.0
python-dotenv>=1.0.0
gtts>=2.5.1
Pillow>=10.2.0
requests>=2.31.0
plotly>=5.18.0
pandas>=2.0.0
numpy>=1.24.0
streamlit-option-menu>=0.3.12
altair>=5.0.0
